Version 2.0.0: Add a friendly syntax, rework serialization and fix bugs
The EmailListener can now be loaded and saved, and can be invoked to 
load it then run it. 
There is now an 'actions' field in the JSON file containing all the 
defined actions.
There are now more accurate error messages when an error occur when 
loading or saving a Filter or EmailListener.
Fixed untested bugs and added more tests, specially some end to end 
tests.